implement frontend performance optimization
