+++ public/manifest.json
@@ -1,0 +1,21 @@
+{
+  "name": "InheritX",
+  "short_name": "InheritX",
+  "description": "Secure and stress-free inheritance planning.",
+  "start_url": "/",
+  "display": "standalone",
+  "background_color": "#161E22",
+  "theme_color": "#33C5E0",
+  "icons": [
+    {
+      "src": "/logo.png",
+      "sizes": "192x192",
+      "type": "image/png"
+    },
+    {
+      "src": "/logo.png",
+      "sizes": "512x512",
+      "type": "image/png"
+    }
+  ]
+}
