+++ public/sw.js
@@ -1,0 +1,1 @@
+if(!self.define){let e,a={};const s=(s,c)=>(s=new URL(s+".js",c).href,a[s]||new Promise(a=>{if("document"in self){const e=document.createElement("script");e.src=s,e.onload=a,document.head.appendChild(e)}else e=s,importScripts(s),a()}).then(()=>{let e=a[s];if(!e)throw new Error(`Module ${s} didn’t register its module`);return e}));self.define=(c,n)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(a[i])return;let t={};const r=e=>s(e,i),d={module:{uri:i},exports:t,require:r};a[i]=Promise.all(c.map(e=>d[e]||r(e))).then(e=>(n(...e),t))}}define(["./workbox-f1770938"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/ArrowUp.png",revision:"e7384b21a1b016ec7f9a0cc0ca8617a2"},{url:"/BarCode.svg",revision:"e5824bbe114865eebb4021076eca2408"},{url:"/Rectangle 1046.svg",revision:"ab6dd3c0cb05a69e36a5ca9379e469e6"},{url:"/Vector (1).svg",revision:"7dcc011e0f109c9be2411b767a19f366"},{url:"/Vector (2).svg",revision:"f7ac9b6b185462b77a8caecc4a19937f"},{url:"/Vector (3).svg",revision:"20daac263176afacfd1472ca6b16e3ad"},{url:"/_next/static/DLnPoM8QEXAEcLxb4_qNu/_buildManifest.js",revision:"e8b26a1719bd9b7e1bbbf62eeb87efbe"},{url:"/_next/static/DLnPoM8QEXAEcLxb4_qNu/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/1955.2c0045e1e62a1785.js",revision:"2c0045e1e62a1785"},{url:"/_next/static/chunks/1966.b8f83f7ba376c962.js",revision:"b8f83f7ba376c962"},{url:"/_next/static/chunks/2002-974ecd76256bdb77.js",revision:"974ecd76256bdb77"},{url:"/_next/static/chunks/3148-3c6b0503a9335a14.js",revision:"3c6b0503a9335a14"},{url:"/_next/static/chunks/3899.3f761f7e3a944e8c.js",revision:"3f761f7e3a944e8c"},{url:"/_next/static/chunks/4286-a6755dd8556d86df.js",revision:"a6755dd8556d86df"},{url:"/_next/static/chunks/4bd1b696-e5d7c65570c947b7.js",revision:"e5d7c65570c947b7"},{url:"/_next/static/chunks/5128.014ddee1031e68f7.js",revision:"014ddee1031e68f7"},{url:"/_next/static/chunks/5513.1c1657ac076417a9.js",revision:"1c1657ac076417a9"},{url:"/_next/static/chunks/5772-a4d889fed696a312.js",revision:"a4d889fed696a312"},{url:"/_next/static/chunks/7703-b7e37237011690b4.js",revision:"b7e37237011690b4"},{url:"/_next/static/chunks/8928-bfa464e02446d9d4.js",revision:"bfa464e02446d9d4"},{url:"/_next/static/chunks/9743-59c0401df25c46c7.js",revision:"59c0401df25c46c7"},{url:"/_next/static/chunks/app/Guidelines/Code-of-ethics/page-30edafa1ee3cd51c.js",revision:"30edafa1ee3cd51c"},{url:"/_next/static/chunks/app/Guidelines/Terms-and-Conditions/page-14e5a071637666ad.js",revision:"14e5a071637666ad"},{url:"/_next/static/chunks/app/Guidelines/page-a24281f5b61019e8.js",revision:"a24281f5b61019e8"},{url:"/_next/static/chunks/app/_global-error/page-de9c4ead561c8302.js",revision:"de9c4ead561c8302"},{url:"/_next/static/chunks/app/_not-found/page-e536bd86654ea9a8.js",revision:"e536bd86654ea9a8"},{url:"/_next/static/chunks/app/about/page-3c7900a9e6c7073d.js",revision:"3c7900a9e6c7073d"},{url:"/_next/static/chunks/app/admin/activity-log/page-4b8fce1b8b7da072.js",revision:"4b8fce1b8b7da072"},{url:"/_next/static/chunks/app/admin/all-plans/page-4eff7333a42c1eaa.js",revision:"4eff7333a42c1eaa"},{url:"/_next/static/chunks/app/admin/insurance-fund/page-3fc29ee4aaf43039.js",revision:"3fc29ee4aaf43039"},{url:"/_next/static/chunks/app/admin/kyc-management/page-a19d0ea4312495ee.js",revision:"a19d0ea4312495ee"},{url:"/_next/static/chunks/app/admin/layout-c648b05298297328.js",revision:"c648b05298297328"},{url:"/_next/static/chunks/app/admin/login/page-120c0545efcb0ac1.js",revision:"120c0545efcb0ac1"},{url:"/_next/static/chunks/app/admin/page-e64c2084db3b391d.js",revision:"e64c2084db3b391d"},{url:"/_next/static/chunks/app/admin/users/page-2f712d90ede1236c.js",revision:"2f712d90ede1236c"},{url:"/_next/static/chunks/app/asset-owner/borrowing/page-188284cc21993b3d.js",revision:"188284cc21993b3d"},{url:"/_next/static/chunks/app/asset-owner/claim/page-47a72ea3ff5c1104.js",revision:"47a72ea3ff5c1104"},{url:"/_next/static/chunks/app/asset-owner/emergency/page-d7e08e576b743c0a.js",revision:"d7e08e576b743c0a"},{url:"/_next/static/chunks/app/asset-owner/inactivity/page-f0ad0393eacaf075.js",revision:"f0ad0393eacaf075"},{url:"/_next/static/chunks/app/asset-owner/layout-0716b11a83afdb21.js",revision:"0716b11a83afdb21"},{url:"/_next/static/chunks/app/asset-owner/lending/page-e283df7c6b2940e9.js",revision:"e283df7c6b2940e9"},{url:"/_next/static/chunks/app/asset-owner/messages/page-9ec0a1fdccdf29c9.js",revision:"9ec0a1fdccdf29c9"},{url:"/_next/static/chunks/app/asset-owner/page-de9c4ead561c8302.js",revision:"de9c4ead561c8302"},{url:"/_next/static/chunks/app/asset-owner/plans/%5Bid%5D/page-a21677513bea02e0.js",revision:"a21677513bea02e0"},{url:"/_next/static/chunks/app/asset-owner/plans/%5Bid%5D/will-documents/page-77e0aab7beec04c2.js",revision:"77e0aab7beec04c2"},{url:"/_next/static/chunks/app/asset-owner/plans/page-01fa4446583a6c5f.js",revision:"01fa4446583a6c5f"},{url:"/_next/static/chunks/app/asset-owner/portfolio/page-de9c4ead561c8302.js",revision:"de9c4ead561c8302"},{url:"/_next/static/chunks/app/asset-owner/security/page-d835383046ec0a87.js",revision:"d835383046ec0a87"},{url:"/_next/static/chunks/app/asset-owner/swap/page-0006f056c4472bc4.js",revision:"0006f056c4472bc4"},{url:"/_next/static/chunks/app/contact/page-1bb120a474ad952e.js",revision:"1bb120a474ad952e"},{url:"/_next/static/chunks/app/faqs/page-6760e796300660dd.js",revision:"6760e796300660dd"},{url:"/_next/static/chunks/app/how-it-works/page-eff9d0e915026d75.js",revision:"eff9d0e915026d75"},{url:"/_next/static/chunks/app/layout-fd544681076edf0a.js",revision:"fd544681076edf0a"},{url:"/_next/static/chunks/app/page-41d29d1fa5d6d6ed.js",revision:"41d29d1fa5d6d6ed"},{url:"/_next/static/chunks/app/privacy-policy/page-41ae9f3f294ac43a.js",revision:"41ae9f3f294ac43a"},{url:"/_next/static/chunks/framework-81b2e59ffe13bb24.js",revision:"81b2e59ffe13bb24"},{url:"/_next/static/chunks/main-1d0d237864fca987.js",revision:"1d0d237864fca987"},{url:"/_next/static/chunks/main-app-1e63e7bbee5c1157.js",revision:"1e63e7bbee5c1157"},{url:"/_next/static/chunks/next/dist/client/components/builtin/app-error-de9c4ead561c8302.js",revision:"de9c4ead561c8302"},{url:"/_next/static/chunks/next/dist/client/components/builtin/forbidden-de9c4ead561c8302.js",revision:"de9c4ead561c8302"},{url:"/_next/static/chunks/next/dist/client/components/builtin/global-error-f7e73f242f568b22.js",revision:"f7e73f242f568b22"},{url:"/_next/static/chunks/next/dist/client/components/builtin/not-found-de9c4ead561c8302.js",revision:"de9c4ead561c8302"},{url:"/_next/static/chunks/next/dist/client/components/builtin/unauthorized-de9c4ead561c8302.js",revision:"de9c4ead561c8302"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-e2411b539ac716ac.js",revision:"e2411b539ac716ac"},{url:"/_next/static/css/29b53fa8ba15bf77.css",revision:"29b53fa8ba15bf77"},{url:"/_next/static/media/747892c23ea88013-s.woff2",revision:"a0761690ccf4441ace5cec893b82d4ab"},{url:"/_next/static/media/7b0b24f36b1a6d0b-s.p.woff2",revision:"98ccc2b7f18991a5126a91ac56fbb1fc"},{url:"/_next/static/media/93f479601ee12b01-s.p.woff2",revision:"da83d5f06d825c5ae65b7cca706cb312"},{url:"/_next/static/media/9610d9e46709d722-s.woff2",revision:"7b7c0ef93df188a852344fc272fc096b"},{url:"/_next/static/media/98848575513c9742-s.woff2",revision:"e2b64ddcb351dbe7397e0da426a8c8d6"},{url:"/_next/static/media/BarCode.6f3b8380.svg",revision:"e5824bbe114865eebb4021076eca2408"},{url:"/_next/static/media/Rectangle 1046.22d9077a.svg",revision:"ab6dd3c0cb05a69e36a5ca9379e469e6"},{url:"/_next/static/media/error.2841a4f2.svg",revision:"c5318c23419bd438a066c484c367111d"},{url:"/_next/static/media/faceID.90432149.svg",revision:"29a422497dfa5da2b4f24bde50e0e9ae"},{url:"/_next/static/media/stuff.ca68dcec.svg",revision:"a077b353caeffbf72b26117dabcfa905"},{url:"/arrowdown.png",revision:"93004d325732438cb55586e9234f6fa6"},{url:"/circle.png",revision:"8568b4587c767957b415f73acbcaf1b6"},{url:"/circle.svg",revision:"6013eb8bd85c2a092e2eb78d4c1214c8"},{url:"/error.svg",revision:"c5318c23419bd438a066c484c367111d"},{url:"/faceID.svg",revision:"29a422497dfa5da2b4f24bde50e0e9ae"},{url:"/faq.png",revision:"1768986e236c6488f4e2f284e9c5ba3e"},{url:"/faq2.png",revision:"289c9c07bce6b00efdc59908ee754ec2"},{url:"/file.svg",revision:"d09f95206c3fa0bb9bd9fefabfd0ea71"},{url:"/globe.svg",revision:"e52c08ce0a255991e9cbf23bd6828555"},{url:"/how it works.svg",revision:"4ff0715c1a98829e4d6f737c33ef9311"},{url:"/logo.png",revision:"b3b731becb3f3ecf6a091bb09ac1bc6e"},{url:"/logo.svg",revision:"7ed8f3757b2c149e8d1db59713acc970"},{url:"/manifest.json",revision:"6efe8d8bf06e567fe9ba9325550e394c"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/stuff.svg",revision:"a077b353caeffbf72b26117dabcfa905"},{url:"/tree.svg",revision:"25abf5538d5397d77d087a6d0b2ec74d"},{url:"/tree2.svg",revision:"88bfbba0fc540e09337a1001a7e3e257"},{url:"/user-avatar.svg",revision:"31cdbbff56e253fdf4ac895fa4e93345"},{url:"/vercel.svg",revision:"c0af2f507b369b085b35ef4bbe3bcf1e"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({response:e})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:2592e3})]}),"GET"),e.registerRoute(/\/_next\/static.+\.js$/i,new e.CacheFirst({cacheName:"next-static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4|webm)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:48,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({sameOrigin:e,url:{pathname:a}})=>!(!e||a.startsWith("/api/auth/callback")||!a.startsWith("/api/")),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({request:e,url:{pathname:a},sameOrigin:s})=>"1"===e.headers.get("RSC")&&"1"===e.headers.get("Next-Router-Prefetch")&&s&&!a.startsWith("/api/"),new e.NetworkFirst({cacheName:"pages-rsc-prefetch",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({request:e,url:{pathname:a},sameOrigin:s})=>"1"===e.headers.get("RSC")&&s&&!a.startsWith("/api/"),new e.NetworkFirst({cacheName:"pages-rsc",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:{pathname:e},sameOrigin:a})=>a&&!e.startsWith("/api/"),new e.NetworkFirst({cacheName:"pages",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({sameOrigin:e})=>!e,new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")});
